Refactor: move reorganize files, switch tab bar delegate for callback.
