+++ Client/Frontend/BraveUX.swift
@@ -11,6 +11,7 @@ struct BraveUX {
     struct TabsBar {
         static let buttonWidth = UIDevice.current.userInterfaceIdiom == .pad ? 40 : 0
         static let height: CGFloat = 29
+        static let minimumWidth: CGFloat =  UIDevice.current.userInterfaceIdiom == .pad ? 180 : 160
     }
 
     struct UrlBar {
